Introduced server and client channels to init
Every process will get a server and a client channel to its core local
init. A server channel is for the process to listen for requests from
init and a client channel is for the process to send requests to init.
Only the server channel is registered on an event handler. Therefore,
the client channel is safe to be used with LMP calls